Fixed project structure for compilation
Changed package name from cz.fi.muni.... to cz.muni.fi....
Fixed typos.
