+++ esports-api/src/main/java/CompetitionFacade.java
@@ -1,0 +1,2 @@
+public interface CompetitionFacade {
+}
