--- esports-api/src/main/java/CompitionFacade.java
@@ -1,2 +1,0 @@
-public interface CompitionFacade {
-}
