--- esports-persistence/src/main/java/cz/fi/muni/pa165/esports/enums/Game.java
@@ -1,8 +1,0 @@
-package cz.fi.muni.pa165.esports.enums;
-/**
- * @author Elena Álvarez
- */
-public enum Game {
-    FIFA,
-    LOL
-}
