--- esports-persistence/src/main/java/cz/fi/muni/pa165/esports/enums/Gender.java
@@ -1,10 +1,0 @@
-package cz.fi.muni.pa165.esports.enums;
-
-/**
- * @author Radovan Tomasik
- */
-public enum Gender {
-    MALE,
-    FEMALE,
-    OTHER
-}
