+++ esports-persistence/src/main/java/cz/muni/fi/pa165/esports/enums/Game.java
@@ -1,0 +1,8 @@
+package cz.muni.fi.pa165.esports.enums;
+/**
+ * @author Elena Álvarez
+ */
+public enum Game {
+    FIFA,
+    LOL
+}
