+++ esports-persistence/src/main/java/cz/muni/fi/pa165/esports/enums/Gender.java
@@ -1,0 +1,10 @@
+package cz.muni.fi.pa165.esports.enums;
+
+/**
+ * @author Radovan Tomasik
+ */
+public enum Gender {
+    MALE,
+    FEMALE,
+    OTHER
+}
