--- esports-service/src/main/java/CompetitionService.java
@@ -1,3 +1,0 @@
-public interface CompetitionService {
-
-}
