--- esports-service/src/main/java/CompetitionServiceImpl.java
@@ -1,2 +1,0 @@
-public class CompetitionServiceImpl implements CompetitionService {
-}
