--- esports-service/src/main/java/MatchRecordService.java
@@ -1,4 +1,0 @@
-public interface MatchRecordService {
-
-
-}
