--- esports-service/src/main/java/MatchRecordServiceImpl.java
@@ -1,2 +1,0 @@
-public class MatchRecordServiceImpl implements MatchRecordService {
-}
