--- esports-service/src/main/java/TeamServiceImp.java
@@ -1,27 +1,0 @@
-import cz.fi.muni.pa165.esports.entity.Competition;
-import cz.fi.muni.pa165.esports.entity.Player;
-
-import java.util.List;
-
-public class TeamServiceImp implements TeamService {
-
-    @Override
-    public boolean applyForCompetion(Competition competition) {
-        return false;
-    }
-
-    @Override
-    public void acceptPlayer(Player player) {
-
-    }
-
-    @Override
-    public List<Player> getPlayers() {
-        return null;
-    }
-
-    @Override
-    public void getTeamStatistics() {
-
-    }
-}
