+++ esports-service/src/main/java/cz/muni/fi/pa165/esports/service/CompetitionService.java
@@ -1,0 +1,5 @@
+package cz.muni.fi.pa165.esports.service;
+
+public interface CompetitionService {
+
+}
