+++ esports-service/src/main/java/cz/muni/fi/pa165/esports/service/MatchRecordService.java
@@ -1,0 +1,6 @@
+package cz.muni.fi.pa165.esports.service;
+
+public interface MatchRecordService {
+
+
+}
