+++ esports-service/src/main/java/cz/muni/fi/pa165/esports/service/MatchRecordServiceImpl.java
@@ -1,0 +1,4 @@
+package cz.muni.fi.pa165.esports.service;
+
+public class MatchRecordServiceImpl implements MatchRecordService {
+}
