+++ pom.xml
@@ -18,7 +18,7 @@
         <relativePath/>
     </parent>
 
-    <groupId>cz.fi.muni.pa165</groupId>
+    <groupId>cz.muni.fi.pa165</groupId>
     <artifactId>esports</artifactId>
     <version>1.0-SNAPSHOT</version>
